+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './object';
 export * from './color';
 export * from './format';
 export * from './platform';
+export * from './url';